{{cookiecutter.repo_name}}/requirements/production.txt
# General
more-itertools~=9.0
cryptography~=38.0

#Flask
Flask~=2.2
flask-login~=0.6
gunicorn~=20.1
pyjwt~=2.6.0
flask-wtf~=1.0

# Database
flask-sqlalchemy~=3.0
flask-migrate~=4.0
psycopg2-binary~=2.9
sqlalchemy~=1.4

# Socket IO
Flask-SocketIO~=5.3
python-engineio~=4.3
python-socketio~=5.7
gevent-websocket~=0.10

# Scipy
numpy~=1.24
sympy~=1.11
scipy~=1.9

# Background
rq~=1.11
redis~=4.4
APScheduler~=3.9

